fix: Changes to the value of items in the list were not notified outside the UI.
The Setter in the ListBinder was not being called.
So, libraries such as PrefsGUI were not receiving notification of changes to the List.
