+++ Packages/RosettaUI/Runtime/Binder/IListItemBinder.cs
@@ -1,0 +1,7 @@
+﻿namespace RosettaUI
+{
+    public interface IListItemBinder : IBinder
+    {
+        int Index { get; set; }
+    }
+}
