+++ Packages/RosettaUI/Runtime/Binder/IListItemBinder.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: eb8d42c9ca9547c09d53d3e266eedb1d
+timeCreated: 1687337690
