+++ Packages/RosettaUI/Runtime/Binder/ListBinder.cs
@@ -10,7 +10,9 @@ namespace RosettaUI
     {
         public static IEnumerable<BinderBase<T>> CreateItemBinders<T>(IList<T> list)
         {
-            return Enumerable.Range(0, list.Count).Select(i => new ListItemBinder<T>(ConstGetter.Create(list), i));
+            // var listBinder = 
+            // return Enumerable.Range(0, list.Count).Select(i => new ListItemBinder<T>(, i));
+            return null;
         }
 
         public static Type GetItemBinderType(Type type)
@@ -118,29 +120,6 @@ namespace RosettaUI
             var itemType = ListUtility.GetItemType(binder.ValueType);
             list = ListUtility.RemoveItemAtLast(list, itemType);
             binder.SetObject(list);
-        }
-    }
-
-    public interface IListItemBinder : IBinder
-    {
-        int Index { get; set; }
-    }
-
-    public class ListItemBinder<T> : ChildBinder<IList<T>, T>, IListItemBinder
-    {
-        public int Index { get; set; }
-        
-        public ListItemBinder(IGetter<IList<T>> listGetter, int index) : 
-            base(new Binder<IList<T>>(listGetter, null))
-        {
-            Index = index;
-        }
-
-        protected override T GetFromParent(IList<T> list) => (0 <= Index && Index < list.Count) ? list[Index] : default;
-        protected override IList<T> SetToParent(IList<T> list, T value)
-        {
-            if ( Index < list.Count ) list[Index] = value;
-            return list; 
         }
     }
 }
