+++ Packages/RosettaUI/Runtime/Binder/ListItemBinder.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 2b89cf34ea564b9781674493e8033c5e
+timeCreated: 1687337712
